Their punctuation is exactly the same, except that their punctuation mark that looks like 7 are encoded separately for each script.
1) in namo tassa, convert double danda to "the 7".
2) in gathas, convert single danda to semicolon and double danda to "the 7".
3) all other single and double dandas convert to "the 7".

The current behavior matches the CSCD, but FS has a hunch that it's not
correct. There is another punctuation mark in both scripts that looks
like the 7 but has a line parallel to it, almost like "71". The Unicode
docs make it seem like the 7 and 71 are similar in behavior to single
and double danda.

git-svn-id: https://tipitaka.org/svn@110 4d745937-c62f-0410-b93d-c932bdf89b14